Provide skip setup option for iota harness (#16284)

+++ iota/test/iris/utils/cleanup.py
@@ -1,9 +1,13 @@
 #! /usr/bin/python3
 import iota.harness.api as api
+from iota.harness.infra.glopts import GlobalOptions as GlobalOptions
 
 # The cleanup step cleans up the previous asicerrord files
 # so that we can start with a clean setup.
 def Main(tc):
+    if GlobalOptions.skip_setup:
+        return api.types.status.SUCCESS
+
     req = api.Trigger_CreateExecuteCommandsRequest(serial = True)
     cmd = "rm /obfl/asicerrord_err.log.*"
     cmd1 = "truncate -s 0 /obfl/asicerrord_err.log"
